removing more  netgame code and implementing some improvements following early fastdoom commits   (aug 19)

+++ r_defs.h
@@ -374,8 +374,9 @@ typedef struct vissprite_s
     struct vissprite_s*	prev;
     struct vissprite_s*	next;
     
-    int			x1;
-    int			x2;
+
+	short x1;
+	short x2;
 
     // for line side calculation
     fixed_t		gx;
@@ -394,7 +395,7 @@ typedef struct vissprite_s
     fixed_t		xiscale;	
 
     fixed_t		texturemid;
-    int			patch;
+	short patch;
 
     // for color translation and shadow draw,
     //  maxbright frames as well
@@ -457,9 +458,9 @@ typedef struct
 {
   fixed_t		height;
   int			picnum;
-  int			lightlevel;
-  int			minx;
-  int			maxx;
+  short lightlevel;
+  short minx;
+  short maxx;
   
   // leave pads for [minx-1]/[maxx+1]
